Ensure `<title>`s don't contain HTML tags.
For a blog post, the `<title>` element will be populated by the post's original title. A post whose title included HTML (e.g. a title specified in Markdown like: `This is a *really* good post`) previously led to tags being nested into the `<title>` element. We now pass the contents of `<title>` through the `striptags` filter to avoid this.